Refactor worker into modular components
